reworking examples on v2 pages into new format,  minor fixes for Button and IconButton
