+++ lib/moon/design/button.ex
@@ -47,8 +47,8 @@ defmodule Moon.Design.Button do
       {=@id}
       class={merge([
         [
-          "flex row justify-center items-center gap-2 relative overflow-hidden active:scale-90 transition-all font-semibold group z-0",
-          "whitespace-nowrap select-none transition duration-200 active:scale-90",
+          "flex row justify-center items-center gap-2 relative overflow-hidden active:scale-90 transition-all duration-200 font-semibold group z-0",
+          "whitespace-nowrap select-none",
           Utils.get_button_size_classes(@size),
           "#{Utils.get_no_icon_padding(@size)}": !@right_icon && !@left_icon,
           "#{Utils.get_right_icon_paddings(@size)}": @right_icon && !@full_width,
@@ -60,7 +60,7 @@ defmodule Moon.Design.Button do
           "bg-hit text-goten": @variant in ["tertiary"],
           "bg-none text-trunks hover:text-bulma": @variant in ["ghost"],
           "w-full": @full_width,
-          "opacity-30 cursor-not-allowed": @disabled,
+          "opacity-30 cursor-not-allowed active:transform-none": @disabled,
           "anim-pulse animate-[pulse2_1.5s_infinite]": @animation == "pulse",
           "bg-chichi text-goten animate-[error_0.82s_cubic-bezier(0.36,0.07,0.19,0.97)_1_both] anim-error":
             @animation == "error"
