+++ lib/moon/design/button/icon_button.ex
@@ -38,15 +38,14 @@ defmodule Moon.Design.Button.IconButton do
       {=@id}
       class={merge([
         [
-          "flex row justify-center items-center gap-2 relative overflow-hidden active:scale-90 transition-all font-semibold group z-0",
-          "select-none transition duration-200 active:scale-90",
+          "flex row justify-center items-center gap-2 relative overflow-hidden active:scale-90 transition-all duration-200 select-none font-semibold group z-0",
           get_icon_button_size_classes(@size),
           "text-goten bg-piccolo": @variant in ["primary"],
           "border border-solid bg-transparent text-bulma border-trunks hover:border-bulma":
             @variant in ["secondary"],
           "bg-hit text-goten": @variant in ["tertiary"],
           "bg-none text-trunks hover:text-bulma": @variant in ["ghost"],
-          "opacity-30 cursor-not-allowed": @disabled,
+          "opacity-30 cursor-not-allowed active:transform-none": @disabled,
           "anim-pulse animate-[pulse2_1.5s_infinite]": @animation == "pulse",
           "bg-chichi text-goten animate-[error_0.82s_cubic-bezier(0.36,0.07,0.19,0.97)_1_both] anim-error":
             @animation == "error"
