+++ lib/moon_web/examples/design/icon_button_example/default.ex
@@ -1,0 +1,24 @@
+defmodule MoonWeb.Examples.Design.IconButtonExample.Default do
+  @moduledoc false
+
+  use Moon.StatelessComponent
+  use MoonWeb, :example
+
+  alias Moon.Design.Button.IconButton
+
+  def render(assigns) do
+    ~F"""
+    <IconButton icon_only="generic_settings" />
+    """
+  end
+
+  def code() do
+    """
+    alias Moon.Design.Button.IconButton
+
+    ...
+
+    <IconButton icon_only="generic_settings" />
+    """
+  end
+end
