+++ lib/moon_web/examples/design/loader_example/default.ex
@@ -1,0 +1,24 @@
+defmodule MoonWeb.Examples.Design.LoaderExample.Default do
+  @moduledoc false
+
+  use Moon.StatelessComponent
+  use MoonWeb, :example
+
+  alias Moon.Design.Loader
+
+  def render(assigns) do
+    ~F"""
+    <Loader />
+    """
+  end
+
+  def code() do
+    """
+    alias Moon.Design.Loader
+
+    ...
+
+    <Loader />
+    """
+  end
+end
