+++ lib/moon_web/examples/design/menu_item_example/as_link.ex
@@ -1,0 +1,24 @@
+defmodule MoonWeb.Examples.Design.MenuItemExample.AsLink do
+  @moduledoc false
+
+  use Moon.StatelessComponent
+  use MoonWeb, :example
+
+  alias Moon.Design.MenuItem
+
+  def render(assigns) do
+    ~F"""
+    <MenuItem as="a">Menu item text</MenuItem>
+    """
+  end
+
+  def code() do
+    """
+    alias Moon.Design.MenuItem
+
+    ...
+
+    <MenuItem as="a">Menu item text</MenuItem>
+    """
+  end
+end
