+++ lib/moon_web/pages/design/loader_page.ex
@@ -3,11 +3,12 @@ defmodule MoonWeb.Pages.Design.LoaderPage do
 
   use MoonWeb, :live_view
 
-  alias Moon.Design.Loader
   alias MoonWeb.Components.Page
   alias MoonWeb.Components.ComponentPageDescription
-  alias MoonWeb.Components.ExampleAndCode
+  alias MoonWeb.Components.ExamplesList
   alias MoonWeb.Components.PropsTable
+
+  alias MoonWeb.Examples.Design.LoaderExample
 
   data(breadcrumbs, :any,
     default: [
@@ -22,25 +23,6 @@ defmodule MoonWeb.Pages.Design.LoaderPage do
     ]
   )
 
-  data(props_info_array, :list,
-    default: [
-      %{
-        :name => 'color',
-        :type => 'string',
-        :required => 'No',
-        :default => 'hit',
-        :description => 'Colour of loader'
-      },
-      %{
-        :name => 'size',
-        :type => '2xs | xs | sm | md | lg',
-        :required => 'No',
-        :default => 'md',
-        :description => 'Size of loader'
-      }
-    ]
-  )
-
   def render(assigns) do
     ~F"""
     <Page {=@theme_name} {=@active_page} {=@breadcrumbs} {=@direction}>
@@ -50,62 +32,32 @@ defmodule MoonWeb.Pages.Design.LoaderPage do
         </p>
       </ComponentPageDescription>
 
-      <ExampleAndCode title="Default" id="loader_1">
-        <:example>
-          <Loader />
-        </:example>
+      <ExamplesList examples={[
+        LoaderExample.Default,
+        LoaderExample.Sizes,
+        LoaderExample.Colours
+      ]} />
 
-        <:code>{loader_1_code()}</:code>
-      </ExampleAndCode>
-
-      <ExampleAndCode title="Sizes" id="loader_2">
-        <:example>
-          <Loader size="2xs" />
-          <Loader size="xs" />
-          <Loader size="sm" />
-          <Loader />
-          <Loader size="lg" />
-        </:example>
-
-        <:code>{loader_2_code()}</:code>
-      </ExampleAndCode>
-
-      <ExampleAndCode title="Colours" id="loader_3">
-        <:example>
-          <Loader color="dodoria" />
-          <Loader color="trunks" />
-          <Loader color="raditz" />
-        </:example>
-
-        <:code>{loader_3_code()}</:code>
-      </ExampleAndCode>
-
-      <PropsTable title="Loader props" data={@props_info_array} />
+      <PropsTable
+        title="Loader props"
+        data={[
+          %{
+            :name => 'color',
+            :type => 'string',
+            :required => 'No',
+            :default => 'hit',
+            :description => 'Colour of loader'
+          },
+          %{
+            :name => 'size',
+            :type => '2xs | xs | sm | md | lg',
+            :required => 'No',
+            :default => 'md',
+            :description => 'Size of loader'
+          }
+        ]}
+      />
     </Page>
-    """
-  end
-
-  def loader_1_code do
-    """
-      <Loader/>
-    """
-  end
-
-  def loader_2_code do
-    """
-      <Loader size="2xs" />
-      <Loader size="xs" />
-      <Loader size="sm" />
-      <Loader />
-      <Loader size="lg" />
-    """
-  end
-
-  def loader_3_code do
-    """
-      <Loader color="dodoria"/>
-      <Loader color="trunks" />
-      <Loader color="raditz" />
     """
   end
 end
